GrupoProduto
GrupoProduto

+++ src/main/java/com/br/gcm/controller/GrupoProdutoController.java
@@ -3,8 +3,10 @@ package com.br.gcm.controller;
 import com.br.gcm.dao.GrupoProdutoDao;
 import com.br.gcm.model.GrupoProduto;
 import com.br.gcm.model.MensagemTransacao;
+import com.br.gcm.model.Usuario;
 import com.br.gcm.service.GrupoProdutoService;
 import com.br.gcm.tag.Pagina;
+import com.br.gcm.util.Rotinas;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -32,11 +34,26 @@ import javax.swing.*;
 public class GrupoProdutoController {
     @Inject private GrupoProdutoDao grupoProdutoDao;
     @Inject private GrupoProdutoService grupoProdutoService;
+    @Inject private Rotinas rotinas;
 
      //Listar
     @RequestMapping(value = "/grupoproduto_lista")
     public String lista(@PageableDefault(size = 10) Pageable pageable, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1101");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+        Boolean novo     = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110101");
+        Boolean editar   = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110102");
+        Boolean deletar  = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110103");
+        Boolean detalhes = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110104");
 
+        model.addAttribute("novo", novo);
+        model.addAttribute("editar", editar);
+        model.addAttribute("deletar", deletar);
+        model.addAttribute("detalhes", detalhes);
         model.addAttribute("grupoproduto_lista", grupoProdutoDao.selectAll_paginado(pageable));
         model.addAttribute("pagina", new Pagina(pageable, grupoProdutoDao.count()));
 
@@ -46,6 +63,13 @@ public class GrupoProdutoController {
     //Deletar
     @RequestMapping(value = "/grupoproduto_deleta/{id}")
     public String deletar(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1103");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         try{
             grupoProdutoService.delete(id);
         }catch(Exception e){
@@ -58,6 +82,13 @@ public class GrupoProdutoController {
     //Nova
     @RequestMapping(value = "/grupoproduto_novo", method = RequestMethod.GET)
     public String novo(ModelMap model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110101");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         GrupoProduto grupoProduto = new GrupoProduto();
         model.addAttribute("grupoproduto", grupoProduto);
         return "grupoproduto_novo";
@@ -79,6 +110,13 @@ public class GrupoProdutoController {
     //Editar
     @RequestMapping(value = "/grupoproduto_editar/{id}", method = RequestMethod.GET)
     public String editar(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110102");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         GrupoProduto grupoProduto = grupoProdutoDao.selectById(id);
         model.addAttribute("grupoproduto", grupoProduto);
         return "grupoproduto_editar";
@@ -99,6 +137,13 @@ public class GrupoProdutoController {
 
     @RequestMapping(value = "/grupoproduto_detalhes/{id}", method = RequestMethod.GET)
     public String detalhes(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "110104");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         GrupoProduto grupoProduto = grupoProdutoDao.selectById(id);
         model.addAttribute("grupoproduto", grupoProduto);
         return "grupoproduto_detalhes";
